// Fortran-JS/api/Joinpoints.ts
type PrivateMapper = {
  "Joinpoint": typeof Joinpoint,
  "Program": typeof Program,
  "FileJp": typeof FileJp,
  "Statement": typeof Statement,
  "Expr": typeof Expr,
  "BinaryOperator": typeof BinaryOperator,
  "ExecutableStatement": typeof ExecutableStatement,
  "ActionStatement": typeof ActionStatement,
  "AssignmentStatement": typeof AssignmentStatement,
  "Designator": typeof Designator,
  "DataRef": typeof DataRef,
  "Literal": typeof Literal,
  "IntLiteral": typeof IntLiteral,
  "StringLiteral": typeof StringLiteral,
  "RealLiteral": typeof RealLiteral,
  "DoStatement": typeof DoStatement,
  "LoopControl": typeof LoopControl,
  "RangeLoopControl": typeof RangeLoopControl,
  "StatementBlock": typeof StatementBlock,
  "Execution": typeof Execution,
  "Specification": typeof Specification,
  "ArraySubscriptExpr": typeof ArraySubscriptExpr,
  "CompilerDirective": typeof CompilerDirective,
  "NameValue": typeof NameValue,
  "OmpConstruct": typeof OmpConstruct,
  "OmpBlockConstruct": typeof OmpBlockConstruct,
  "OmpLoopConstruct": typeof OmpLoopConstruct,
  "OmpClause": typeof OmpClause,
  "OmpDataSharingClause": typeof OmpDataSharingClause,
  "OmpReductionClause": typeof OmpReductionClause,
  "OmpOrderedClause": typeof OmpOrderedClause,
  "UseStatement": typeof UseStatement,
  "ProgramUnit": typeof ProgramUnit,
  "MainProgram": typeof MainProgram,
  "Subroutine": typeof Subroutine,
  "SpecificationStatement": typeof SpecificationStatement,
  "TypeDeclarationStatement": typeof TypeDeclarationStatement,
  "AttributeSpecifier": typeof AttributeSpecifier,
  "KeywordAttributeSpecifier": typeof KeywordAttributeSpecifier,
  "ParameterKeyword": typeof ParameterKeyword,
  "FortranDecl": typeof FortranDecl,
  "EntityDecl": typeof EntityDecl,
  "Initialization": typeof Initialization,
  "ExprInitialization": typeof ExprInitialization,
  "IfStatement": typeof IfStatement,
  "IfConstruct": typeof IfConstruct,
  "IfThenBlock": typeof IfThenBlock,
  "IfThenStatement": typeof IfThenStatement,
  "ElseIfBlock": typeof ElseIfBlock,
  "ElseIfStatement": typeof ElseIfStatement,
  "ElseBlock": typeof ElseBlock,
  "SectionSubscript": typeof SectionSubscript,
  "Subscript": typeof Subscript,
};

type DefaultAttributeMap = {
  FileJp: "name",
}
// ...
export class Expr extends Joinpoint {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
}

  /**
   * Represents a binary operation
   */
export class BinaryOperator extends Expr {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
  get kind(): BinaryOperatorKind { return wrapJoinPoint(this._javaObject.kind()) }
  get left(): Expr { return wrapJoinPoint(this._javaObject.left()) }
  set left(value: Expr) { this._javaObject.setLeft(unwrapJoinPoint(value)); }
  get right(): Expr { return wrapJoinPoint(this._javaObject.right()) }
  set right(value: Expr) { this._javaObject.setRight(unwrapJoinPoint(value)); }
  /**
   * Sets the left-hand side of the operation
   */
  setLeft(lhs: Expr): void { return wrapJoinPoint(this._javaObject.setLeft(unwrapJoinPoint(lhs))); }
  /**
   * Sets the right-hand side of the operation
   */
  setRight(rhs: Expr): void { return wrapJoinPoint(this._javaObject.setRight(unwrapJoinPoint(rhs))); }
}

  /**
   * Represents an executable statement
   */
export class ExecutableStatement extends Statement {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
}
// ...
export class Designator extends Expr {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
}

  /**
   * Represents a data reference
   */
export class DataRef extends Designator {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
  get name(): string { return wrapJoinPoint(this._javaObject.name()) }
}

  /**
   * Represents a literal
   */
export class Literal extends Expr {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
  get literal(): string { return wrapJoinPoint(this._javaObject.literal()) }
}

export class IntLiteral extends Literal {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
}
// ...
export class DoStatement extends ExecutableStatement {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
  get body(): Execution { return wrapJoinPoint(this._javaObject.body()) }
  get control(): LoopControl { return wrapJoinPoint(this._javaObject.control()) }
  get kind(): DoKind { return wrapJoinPoint(this._javaObject.kind()) }
  /**
   * Performs a copy of the do statement, including its controls, but not the body
   */
  copyScope(): DoStatement { return wrapJoinPoint(this._javaObject.copyScope()); }
  /**
   * Returns true if the given do statement has the same loop control as this do statement
   */
  sameScope(loop: DoStatement): boolean { return wrapJoinPoint(this._javaObject.sameScope(unwrapJoinPoint(loop))); }
}

  /**
   * Represents the loop control structure, with specialized subclasses for different loop kinds
   */
export class LoopControl extends Joinpoint {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
}

export class RangeLoopControl extends LoopControl {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
  get lower(): Expr { return wrapJoinPoint(this._javaObject.lower()) }
  get step(): Expr { return wrapJoinPoint(this._javaObject.step()) }
  set step(value: Expr) { this._javaObject.setStep(unwrapJoinPoint(value)); }
  get upper(): Expr { return wrapJoinPoint(this._javaObject.upper()) }
  set upper(value: Expr) { this._javaObject.setUpper(unwrapJoinPoint(value)); }
  get var(): DataRef { return wrapJoinPoint(this._javaObject.var()) }
  setStep(step: Expr): void { return wrapJoinPoint(this._javaObject.setStep(unwrapJoinPoint(step))); }
  setUpper(upper: Expr): void { return wrapJoinPoint(this._javaObject.setUpper(unwrapJoinPoint(upper))); }
}

export class StatementBlock extends Joinpoint {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
  get stmts(): Statement[] { return wrapJoinPoint(this._javaObject.stmts()) }
}

export class Execution extends StatementBlock {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
  get executableStmts(): ExecutableStatement[] { return wrapJoinPoint(this._javaObject.executableStmts()) }
  insertBegin(stmt: ExecutableStatement): void { return wrapJoinPoint(this._javaObject.insertBegin(unwrapJoinPoint(stmt))); }
  insertEnd(stmt: ExecutableStatement): void { return wrapJoinPoint(this._javaObject.insertEnd(unwrapJoinPoint(stmt))); }
}
// ...
export class OmpBlockConstruct extends OmpConstruct {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
  setBody(body: Execution): void { return wrapJoinPoint(this._javaObject.setBody(unwrapJoinPoint(body))); }
}

  /**
   * Represents an OpenMP loop construct (such as do or do parallel)
   */
export class OmpLoopConstruct extends OmpConstruct {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
  setLoop(loop: DoStatement): void { return wrapJoinPoint(this._javaObject.setLoop(unwrapJoinPoint(loop))); }
}

  /**
   * Represents an OpenMP clause
   */
export class OmpClause extends Joinpoint {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
}

  /**
   * Represents an OpenMP datasharing clause (public, private, ...)
   */
export class OmpDataSharingClause extends OmpClause {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
}

export class OmpReductionClause extends OmpClause {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
}

export class OmpOrderedClause extends OmpClause {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
}

export class UseStatement extends Statement {
  /**
   * @internal
   */
  static readonly _defaultAttributeInfo: {readonly map?: DefaultAttributeMap, readonly name: string | null, readonly type?: PrivateMapper, readonly jpMapper?: typeof JoinpointMapper} = {
    name: null,
  };
  get moduleName(): string { return wrapJoinPoint(this._javaObject.moduleName()) }
}
// ...
export const DoKind = {
  while: "while",
  range: "range",
  concurrent: "concurrent",
} as const;
export type DoKind = typeof DoKind[keyof typeof DoKind];
// ...
export type BinaryOperatorKind = typeof BinaryOperatorKind[keyof typeof BinaryOperatorKind];

const JoinpointMapper = {
  joinpoint: Joinpoint,
  program: Program,
  file: FileJp,
  statement: Statement,
  expr: Expr,
  binaryOperator: BinaryOperator,
  executableStatement: ExecutableStatement,
  actionStatement: ActionStatement,
  assignmentStatement: AssignmentStatement,
  designator: Designator,
  dataRef: DataRef,
  literal: Literal,
  intLiteral: IntLiteral,
  stringLiteral: StringLiteral,
  realLiteral: RealLiteral,
  doStatement: DoStatement,
  loopControl: LoopControl,
  rangeLoopControl: RangeLoopControl,
  statementBlock: StatementBlock,
  execution: Execution,
  specification: Specification,
  arraySubscriptExpr: ArraySubscriptExpr,
  compilerDirective: CompilerDirective,
  nameValue: NameValue,
  ompConstruct: OmpConstruct,
  ompBlockConstruct: OmpBlockConstruct,
  ompLoopConstruct: OmpLoopConstruct,
  ompClause: OmpClause,
  ompDataSharingClause: OmpDataSharingClause,
  ompReductionClause: OmpReductionClause,
  ompOrderedClause: OmpOrderedClause,
  useStatement: UseStatement,
  programUnit: ProgramUnit,
  mainProgram: MainProgram,
  subroutine: Subroutine,
  specificationStatement: SpecificationStatement,
  typeDeclarationStatement: TypeDeclarationStatement,
  attributeSpecifier: AttributeSpecifier,
  keywordAttributeSpecifier: KeywordAttributeSpecifier,
  parameterKeyword: ParameterKeyword,
  fortranDecl: FortranDecl,
  entityDecl: EntityDecl,
  initialization: Initialization,
  exprInitialization: ExprInitialization,
  ifStatement: IfStatement,
  ifConstruct: IfConstruct,
  ifThenBlock: IfThenBlock,
  ifThenStatement: IfThenStatement,
  elseIfBlock: ElseIfBlock,
  elseIfStatement: ElseIfStatement,
  elseBlock: ElseBlock,
  sectionSubscript: SectionSubscript,
  subscript: Subscript,
};